LayoutTests/imported/w3c/web-platform-tests/fetch/api/redirect/redirect-count-cross-origin-worker-expected.txt


FAIL Redirect 20 times, last redirection to cross-origin assert_equals: Response's status is 200 expected 200 but got 404
FAIL Redirect 21 times, last redirection to cross-origin assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Redirect 20 times, going to cross-origin after 10 assert_equals: Response's status is 200 expected 200 but got 404
FAIL Redirect 21 times, going to cross-origin after 10 assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Redirect 20 times, last redirection to cross-origin with preflight assert_equals: Response's status is 200 expected 200 but got 404
FAIL Redirect 21 times, last redirection to cross-origin with preflight assert_unreached: Should have rejected: undefined Reached unreachable code
FAIL Redirect 20 times, going to cross-origin after 10 with preflight assert_equals: Response's status is 200 expected 200 but got 404
FAIL Redirect 21 times, going to cross-origin after 10 with preflight assert_unreached: Should have rejected: undefined Reached unreachable code
